+++ FastAnimations/Handlers/FishingHandler.cs
@@ -33,7 +33,7 @@ namespace Pathoschild.Stardew.FastAnimations.Handlers
         /// <param name="playerAnimationID">The player's current animation ID.</param>
         public override void Update(int playerAnimationID)
         {
-            this.SpeedUpPlayer(this.Multiplier);
+            this.SpeedUpPlayer();
         }
     }
 }