+++ FastAnimations/Handlers/TreeFallingHandler.cs
@@ -74,13 +74,11 @@ namespace Pathoschild.Stardew.FastAnimations.Handlers
         /// <param name="playerAnimationID">The player's current animation ID.</param>
         public override void Update(int playerAnimationID)
         {
+            GameTime gameTime = Game1.currentGameTime;
+
+            int skips = this.GetSkipsThisTick();
             foreach (var pair in this.GetFallingTrees())
-            {
-                // speed up animation
-                GameTime gameTime = Game1.currentGameTime;
-                for (int i = 1; i < this.Multiplier; i++)
-                    pair.Value.tickUpdate(gameTime, pair.Key, Game1.currentLocation);
-            }
+                this.ApplySkips(skips, () => pair.Value.tickUpdate(gameTime, pair.Key, Game1.currentLocation));
         }